moved to MainActivityNavigation from view sink file
